split console part for further testing
