+++ src/main/java/META-INF/MANIFEST.MF
@@ -1,3 +1,3 @@
 Manifest-Version: 1.0
-Main-Class: ru.spbau.sazanovich.nikita.ConsoleApp
+Main-Class: ru.spbau.sazanovich.nikita.console.ConsoleApp
