+++ src/main/java/ru/spbau/sazanovich/nikita/console/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Provides classes to access a MyGit repository from a terminal.
+ */
+package ru.spbau.sazanovich.nikita.console;
