--- src/main/java/ru/spbau/sazanovich/nikita/package-info.java
@@ -1,4 +1,0 @@
-/**
- * Contains console interface's class.
- */
-package ru.spbau.sazanovich.nikita;
